create code to request movie name from the API when search from the navbar.
create dom function to send movie name from search bar to the second function.
create a function to send a get request to the server with the movie name.
handle the prev request.
create a function in the handler that will send another get request to the movies API with movie name and API_KEY from the server.
Relates #35